Refactor scenario creation into modular components
Extracted CreateScenarioForm, ScenarioList, and CreateScenarioProgressModal into separate files for better modularity and maintainability. Moved useCreateScenario hook logic into create-scenario-forms.js. Updated CreateScenario.jsx to use new components, simplifying the main component and improving code organization.
