+++ src/features/scenario/hooks/create-scenario-forms.js
@@ -32,3 +32,47 @@ export const useFetchWeather = () => {
 
   return weather;
 };
+
+export const useCreateScenario = (projectPath, { onSuccess }) => {
+  const [formData, setFormData] = useState({});
+  const [fetching, setFetching] = useState(false);
+  const [error, setError] = useState();
+
+  const createScenario = async (data) => {
+    setError(null);
+    setFetching(true);
+
+    try {
+      const formattedData = {};
+
+      Object.keys(data).forEach((key) => {
+        // Convert objects to strings
+        if (typeof data[key] === 'object' && !(data[key] instanceof File)) {
+          formattedData[key] = JSON.stringify(data[key]);
+        } else {
+          formattedData[key] = data[key];
+        }
+      });
+      formattedData['project'] = projectPath;
+
+      const response = await apiClient.postForm(
+        `/api/project/scenario/v2`,
+        formattedData,
+      );
+      onSuccess?.(response.data);
+    } catch (error) {
+      console.log(error?.response?.data || error);
+      setError(error?.response?.data || error);
+    } finally {
+      setFetching(false);
+    }
+  };
+
+  useEffect(() => {
+    if (formData?.scenario_name && projectPath) {
+      createScenario(formData);
+    }
+  }, [formData, projectPath]);
+
+  return { setFormData, fetching, error };
+};
